+++ kernel/memory.c
@@ -2,8 +2,8 @@
 #include <kernel/sections.h>
 
 uint64_t full_ram;
-uint32_t usable_ram;
-uint32_t last_phys_address;
+uintptr_t usable_ram;
+uintptr_t last_phys_address;
 memory_area_t memory_areas[MEMORY_AREAS_SIZE];
 static memory_area_t* current = memory_areas;
